# replit/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateSessionBody {
  userAgent?: string;
}

export interface Session {
  id: string;
  variant: string;
  /** 0=started, 1=viewed image, 2=answered free text, 3=completed */
  currentStep: number;
  freeTextAnswer: string | null;
  multipleChoiceAnswer: string | null;
  /** Randomized options shown to this user */
  multipleChoiceOptions: string[];
  createdAt: string;
  completedAt: string | null;
}

export interface SubmitStepBody {
  /** 1=viewed image, 2=free text answer, 3=multiple choice answer */
  step: number;
  answer?: string | null;
}

export interface SessionResult {
  id: string;
  variant: string;
  currentStep: number;
  freeTextAnswer: string | null;
  multipleChoiceAnswer: string | null;
  multipleChoiceOptions: string[];
  createdAt: string;
  completedAt: string | null;
}

export interface ResultsResponse {
  sessions: SessionResult[];
  total: number;
}

export interface VariantStats {
  variant: string;
  totalStarted: number;
  viewedImage: number;
  answeredFreeText: number;
  completed: number;
}

export interface FunnelStats {
  totalStarted: number;
  viewedImage: number;
  answeredFreeText: number;
  completed: number;
  byVariant: VariantStats[];
}

export interface ErrorResponse {
  error: string;
}