# replit/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Creates a new anonymous session and assigns a random variant image
 * @summary Create a new experiment session
 */
export const CreateSessionBody = zod.object({
  userAgent: zod.string().optional(),
});

/**
 * @summary Get session
 */
export const GetSessionParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetSessionResponse = zod.object({
  id: zod.string(),
  variant: zod.string(),
  currentStep: zod
    .number()
    .describe("0=started, 1=viewed image, 2=answered free text, 3=completed"),
  freeTextAnswer: zod.string().nullable(),
  multipleChoiceAnswer: zod.string().nullable(),
  multipleChoiceOptions: zod
    .array(zod.string())
    .describe("Randomized options shown to this user"),
  createdAt: zod.coerce.date(),
  completedAt: zod.coerce.date().nullable(),
});

/**
 * @summary Submit an answer for the current step
 */
export const SubmitStepParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const SubmitStepBody = zod.object({
  step: zod
    .number()
    .describe("1=viewed image, 2=free text answer, 3=multiple choice answer"),
  answer: zod.string().nullish(),
});

export const SubmitStepResponse = zod.object({
  id: zod.string(),
  variant: zod.string(),
  currentStep: zod
    .number()
    .describe("0=started, 1=viewed image, 2=answered free text, 3=completed"),
  freeTextAnswer: zod.string().nullable(),
  multipleChoiceAnswer: zod.string().nullable(),
  multipleChoiceOptions: zod
    .array(zod.string())
    .describe("Randomized options shown to this user"),
  createdAt: zod.coerce.date(),
  completedAt: zod.coerce.date().nullable(),
});

/**
 * @summary Get all experiment results for download
 */
export const GetResultsResponse = zod.object({
  sessions: zod.array(
    zod.object({
      id: zod.string(),
      variant: zod.string(),
      currentStep: zod.number(),
      freeTextAnswer: zod.string().nullable(),
      multipleChoiceAnswer: zod.string().nullable(),
      multipleChoiceOptions: zod.array(zod.string()),
      createdAt: zod.coerce.date(),
      completedAt: zod.coerce.date().nullable(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get funnel drop-off statistics
 */
export const GetFunnelStatsResponse = zod.object({
  totalStarted: zod.number(),
  viewedImage: zod.number(),
  answeredFreeText: zod.number(),
  completed: zod.number(),
  byVariant: zod.array(
    zod.object({
      variant: zod.string(),
      totalStarted: zod.number(),
      viewedImage: zod.number(),
      answeredFreeText: zod.number(),
      completed: zod.number(),
    }),
  ),
});